[FIX] Move delete handling to helper copy_db
After we have written content to metafile, we will remove that content from DB
(or at least parts of what was written). This is done for both conn and gps,
using a single statement. Move this handling to the helper copy_cb-function,
instead of having close to duplicate code in conn and gps writers.

DELETE is done after link/unlink since we don't want to risk loosing any data.
The backend can handle duplicate entries (they are ignored) and it is more
likely that we will run out of inodes than space.

Also, improve the different select queries so that the results are ordered by
timestamp.