[FEATURE] Support correcting timestamps
When analyzing the data we have collected from some devices, we see some
interesting clock shifts. The hardware clock on the router seems to get written
to at some infrequent interval, causing the time at boot to be at some undefined
point in time (not epoch basically). If it takes some time before ntp gets a
fix, the timestamp on metadata will be very wrong causing all kinds of weird
behavior. For example, we have no way of knowing when an event happened.

Solve this by saying that all timestamps less than a certain time (set to
sometime this evening) are invalid and wait until we get a proper timestamp.
I.e., a timestamp that is larger than the given time. This is a valid
assumption, as timestamps are never meant to move backwars in time. At least not
by such a large margin.

We assume that the time on machines without a good hardware clock/battery will
be set to epoch on each boot, in order to simplify the code slightly. When we
get a proper timestamp, we get the time of day and uptime. Since each timestamp
is offset from 0, we update timestamps to be (gettimeofday() - uptime) +
timestamp. The first gives us an estimate for when the actual boot time was, and
timestamps are offsets from the actual boot time.