[FEATURE] Add support for reading a system session id
A system session id is useful in the case when the clock can't be trusted, or
you risk having multiple sessions with the same l3/l4 session ID. This is the
case when for example timer is reset to epoch or some other fixed point in time.

The structure of the file is expected to be "%llu %llu", where the first value
is the id and the second is a multiplier (in the unlikely event that an uint64
will wrap). The functionality is only used by the sqlite writer for now.