[FEATURE] Export GPS updates
The GPS sqlite writer was just a quick hack for getting GPS data before the
Nornet database was ready. Repurpose the writer to serve Monroe instead by
writing the last known GPS position to the local and remote databases. In order
to prevent GPS spam, write at most one position every 1 second.

This information will be used to show if GPS works or not in the frontend.