[FIX] Try to read session id on application start
If metdata exporter crashes, we will fail to update session id on update message
because the constraint is invalid. If an update message is the first to arrive,
we will insert it in the db with session id (0,0). When we update, this might
cause a conflict with the update message that is already present.

The proper fix to this would be to adjust the constraint in the update table.
However, we don't have time to do that now. This scenario can only happen if
metdata exporter crashes. An ok temporary fix is to (try to) read session id at
the start of metadata exporter. If we have any update in database, it should
already be valid. We can crash between inserting update and reading session id
first time, but that is very unlikely.