Delete old update messages
As a final metadata optimisation, delete any old update mesages left in database
when we make first dump. Old is defined as not having been update for 30
minutes. If metadata exporter for some reason is stopped for a long duration of
time, this will cause some redundant fake mode/quality updates. However, we can
deal with those in the backend (easy to see if we change mode to a mode we are
already in, for example).

Implementing this feature was fairly easy, but we changed the delete parameter
to helper-function used for copying db to a callback. This was required because
we might need to run two delete queries for conn events.