[FIX] Make backwards DB-compatibility work
When I reviewed and approved the changes for switching from JSON to SQL,
I did not review the changes made to the SQL-generation code closely
enough. The SQL generated was not compatible with the version of the
databased used in Monroe/NNE.

Fortunately, the fix is easy and as Monroe/NNE as the only SQL-users
left, the other DB-code can be removed.