sqlite: Improve boot time handling
Until now we have assumed that clock will always either be correct on
boot or be set to epoch. This assumption no longer holds, some of the
devices we use have an RTC that seems very hard to disabled. Instead of
spending time disabling the RTC, we can instead remove the assumption
and this commit is the first step.

We store events in the database also before we have a valid timestamp,
and then update the timestamp once we have a valid fix. Since we assume
that system time starts from zero (this code is not run if we have
assumed that timestamp is already valid), what is stored in the database
is offset (in time) of the event from the system boot time. When we
receive a valid fix, we then calculate the real boot time (time_now -
uptime) and then add the real boot time to the offsets we already have.

When we remove the assumption about zero or valid, the computation
becomes a bit more complicated. The values stored in the database before
we have a valid fix is no longer necessarily an offset from zero, but
instead a valid timestamp. We therefore need to calculate the boot time
relative to system clock when data exporter starts.

When we get a valid timestamp, we need to find the difference between
the previously mentioned real boot time and the original boot time. This
is the offset that will be added to all the timestamps in the database
(if any).

We still have one assumption, and this that the inital time at boot will
always be less than the correct time and that one second is one second.
We have some devices where time seems random (hello 2084, 2102 and
2076), but those devices we can force to epoch. If we get a device with
forcing fails, we need to make some more adjustmenets to this code.
However, I consider that pre-mature optimization and having a very
negative outlook on thins (and it is Friday afternoon ...).