sqlite: Read time after uptime
We should read time after we have read uptime. The reason is that if we
for example read time at sec.999, then second will have wrapped when we
read uptime. While we might risk that the same happens with the new
order of operation, it is easier to work with a larger timestamp than a
large uptime. For example, with a larger timestamp, we risk having one
second as offset to the original boot time. With a small timestamp, we
risk having -1 second as offset and that is slightly harder to deal
with.

The boot time is meant as an estimation anyway, so 100% accurate value
is not needed.