sqlite: Remove debug output, finish epoch ass.
Remove some debug out, as the last commit of removing the epoch
assumption.

There is no need to worry about the time. The issue I tried to solve,
was that for example the real boot time could be one second less than
the original boot time. This can happen if for example the uptime and
timeofday is read at different seconds. One second is not important for
us, there is a higher latency in when we discover that a network is down
for example.

Our goal with the DB calculation is to shift the time offset to that of
the boot time. This is done by doing (timestamp - orig_boot_time) +
real_boot_time, for all timestamps less than real_boot_time. If
real_boot_time is less than orig_boot_time, then the statement has no
effect. This is ok, as we input absolute time in the database.

The code is meant to handle large differences in time, for example from
epoch or some default rtc value. And that works fine.