sqlite: Improve boot time estimation
Our old boot time estimation algorithm has served us well, but started
to show issues when faced with faster devices than data_exporter has
normally been run on. The problem was our use of multiple clock sources,
and more specifically, reading uptime from file.

What could happen was that uptime would increase, but the clock used as
source for gettimeofday() would not. On devices with a (working) RTC or
on restarts of data exporter, this timer-behavior caused our estimate
for the real boot time to be less than the original estimate. When we
detected a clock that has moved backwards, we stop exporting metadata.

This commit improves how we make the estimations by reducing the number
of clock sources. Uptime is only read when we estimate the original boot
time, and then we use CLOCK_MONOTONIC_RAW to estimate the amount of time
that has passed when we estimate the real boot time. Instead of using
gettimeofday() to read the current wall time, we use clock_gettime().

Using clock_gettime() ensures that the real time as well as elapsed time
all come from the same source. We also read the wall time after we read
MONOTONIC_RAW, to avoid the elapsed time being "larger" than the real
time. Worst-case, what can happen with the new approach, is that the
timers have not changed and the diff is 0.